move useChainMaintenanceCommands
