+++ app/src/organisms/ModuleWizardFlows/index.tsx
@@ -17,10 +17,7 @@ import {
 import { getTopPortalEl } from '../../App/portal'
 import { WizardHeader } from '../../molecules/WizardHeader'
 import { useAttachedPipettesFromInstrumentsQuery } from '../../organisms/Devices/hooks'
-import {
-  useChainMaintenanceCommands,
-  useCreateTargetedMaintenanceRunMutation,
-} from '../../resources/runs'
+import { useCreateTargetedMaintenanceRunMutation } from '../../resources/runs'
 import { getIsOnDevice } from '../../redux/config'
 import {
   SimpleWizardBody,
@@ -35,7 +32,10 @@ import { SelectLocation } from './SelectLocation'
 import { Success } from './Success'
 import { DetachProbe } from './DetachProbe'
 import { useNotifyDeckConfigurationQuery } from '../../resources/deck_configuration'
-import { useNotifyCurrentMaintenanceRun } from '../../resources/maintenance_runs'
+import {
+  useChainMaintenanceCommands,
+  useNotifyCurrentMaintenanceRun,
+} from '../../resources/maintenance_runs'
 
 import type { AttachedModule, CommandData } from '@opentrons/api-client'
 import { RUN_STATUS_FAILED } from '@opentrons/api-client'
