+++ app/src/resources/maintenance_runs/index.ts
@@ -1,1 +1,2 @@
 export * from './useNotifyCurrentMaintenanceRun'
+export * from './hooks'
